requirements.txt
streamlit
Pillow
openai
